+++ sdk/python/kfp/dsl/metrics_utils_test.py
@@ -26,6 +26,7 @@ from kfp.dsl import metrics_utils
 
 from google.protobuf import json_format
 
+
 class MetricsUtilsTest(unittest.TestCase):
 
   def test_confusion_matrix(self):
@@ -33,19 +34,22 @@ class MetricsUtilsTest(unittest.TestCase):
     conf_matrix.set_categories(['dog', 'cat', 'horses'])
     conf_matrix.log_row('dog', [2, 6, 0])
     conf_matrix.log_cell('cat', 'dog', 3)
-    with open(os.path.join(os.path.dirname(__file__),
-        'test_data', 'expected_confusion_matrix.json')) as json_file:
+    with open(
+        os.path.join(
+            os.path.dirname(__file__), 'test_data',
+            'expected_confusion_matrix.json')) as json_file:
       expected_json = json.load(json_file)
       self.assertEqual(expected_json, conf_matrix.get_metrics())
 
   def test_bulkload_confusion_matrix(self):
     conf_matrix = metrics_utils.ConfusionMatrix()
-    conf_matrix.load_matrix(['dog', 'cat', 'horses'], [
-      [2, 6, 0], [3, 5,6], [5,7,8]])
+    conf_matrix.load_matrix(['dog', 'cat', 'horses'],
+                            [[2, 6, 0], [3, 5, 6], [5, 7, 8]])
 
-    with open(os.path.join(os.path.dirname(__file__),
-        'test_data',
-        'expected_bulk_loaded_confusion_matrix.json')) as json_file:
+    with open(
+        os.path.join(
+            os.path.dirname(__file__), 'test_data',
+            'expected_bulk_loaded_confusion_matrix.json')) as json_file:
       expected_json = json.load(json_file)
       self.assertEqual(expected_json, conf_matrix.get_metrics())
 
@@ -55,11 +59,12 @@ class MetricsUtilsTest(unittest.TestCase):
     confid_metrics.recall = 24.5
     confid_metrics.falsePositiveRate = 98.4
     expected_dict = {
-      'confidenceThreshold': 24.3,
-      'recall': 24.5,
-      'falsePositiveRate': 98.4
+        'confidenceThreshold': 24.3,
+        'recall': 24.5,
+        'falsePositiveRate': 98.4
     }
     self.assertEqual(expected_dict, confid_metrics.get_metrics())
+
 
 if __name__ == '__main__':
   unittest.main()
