--- sdk/python/tests/dsl/test_data/expected_bulk_loaded_classification_metrics.json
@@ -1,54 +1,0 @@
-{
-  "metadata": {
-    "confidenceMetrics": [
-      {
-        "confidenceThreshold": 53.6,
-        "falsePositiveRate": 85.1,
-        "recall": 52.6
-      },
-      {
-        "confidenceThreshold": 53.6,
-        "falsePositiveRate": 85.1,
-        "recall": 52.6
-      },
-      {
-        "confidenceThreshold": 53.6,
-        "falsePositiveRate": 85.1,
-        "recall": 52.6
-      }
-    ],
-    "confusionMatrix": {
-      "annotationSpecs": [
-        {
-          "displayName": "dog"
-        },
-        {
-          "displayName": "cat"
-        },
-        {
-          "displayName": "horses"
-        }
-      ],
-      "row": [
-        [
-          2.0,
-          6.0,
-          0.0
-        ],
-        [
-          3.0,
-          5.0,
-          6.0
-        ],
-        [
-          5.0,
-          7.0,
-          8.0
-        ]
-      ]
-    }
-  },
-  "type": {
-    "instanceSchema": "title: system.ClassificationMetrics\ntype: object\nproperties:\n  auPrc:\n    type: number\n    format: float\n  auRoc:\n    type: number\n    format: float\n  logLoss:\n    type: number\n    format: float\n  confidenceMetrics:\n    type: array\n    items:\n      type: object\n      properties:\n        confidenceThreshold:\n          type: number\n          format: float\n        maxPredictions:\n          type: integer\n          format: int32\n        recall:\n          type: number\n          format: float\n        precision:\n          type: number\n          format: float\n        falsePositiveRate:\n          type: number\n          format: float\n        f1Score:\n          type: number\n          format: float\n        recallAt1:\n          type: number\n          format: float\n        precisionAt1:\n          type: number\n          format: float\n        falsePositiveRateAt1:\n          type: number\n          format: float\n        f1ScoreAt1:\n          type: number\n          format: float\n        truePositiveCount:\n          type: integer\n          format: int64\n        falsePositiveCount:\n          type: integer\n          format: int64\n        falseNegativeCount:\n          type: integer\n          format: int64\n        trueNegativeCount:\n          type: integer\n          format: int64\n  confusionMatrix:\n    type: object\n    properties:\n      annotationSpecs:\n        type: array\n        items:\n          type: object\n          properties:\n            id:\n              type: string\n            displayName:\n              type: string\n      rows:\n        type: array\n        items:\n          type: object\n          properties:\n            row:\n              type: array\n              items:\n                type: integer\n                format: int64\n"
-  }
-}
