--- sdk/python/tests/dsl/test_data/expected_complete_artifact.json
@@ -1,16 +1,0 @@
-{
-    "metadata": {
-      "custom_arr": [
-        2.0,
-        4.0
-      ],
-      "custom_property": 4.0,
-      "property1": 45.5,
-      "property2": "test string"
-    },
-    "name": "test_artifact",
-    "type": {
-      "instanceSchema": "title: TestArtifact\ntype: object\nproperties:\n  property1:\n    type: number\n    format: float\n    description: >\n      A simple number property.\n  property2:\n    type: string\n    description: >\n      A simple string property.\n"
-    },
-    "uri": "gcs://some_bucket/some_object"
-}
