--- sdk/python/tests/dsl/test_data/expected_dataset_artifact.json
@@ -1,10 +1,0 @@
-{
-    "metadata": {
-      "container_format": "TFRecord",
-      "custom_field": 2.0,
-      "payload_format": "JSON"
-    },
-    "type": {
-      "instanceSchema": "title: kfp.Dataset\ntype: object\nproperties:\n  payload_format:\n    type: string\n  container_format:\n    type: string\n"
-    }
-}
