--- sdk/python/tests/dsl/test_data/expected_metrics.json
@@ -1,12 +1,0 @@
-{
-  "metadata": {
-    "accuracy": 98.4,
-    "custom_metric1": 24.0,
-    "custom_metric2": 24.0,
-    "mean_squared_error": 32.1
-  },
-  "name": "test_artifact",
-  "type": {
-    "instanceSchema": "title: system.Metrics\ntype: object\nproperties:\n  accuracy:\n    type: number\n  precision:\n    type: number\n  recall:\n    type: number\n  f1score:\n    type: number\n  mean_absolute_error:\n    type: number\n  mean_squared_error:\n    type: number\n"
-  }
-}
