--- sdk/python/tests/dsl/test_data/expected_model_artifact.json
@@ -1,17 +1,0 @@
-{
-    "metadata": {
-      "custom_field": {
-        "custom_field_1": 1.0,
-        "custom_field_2": [
-          1.0,
-          3.0,
-          4.0
-        ]
-      },
-      "framework": "TFX",
-      "framework_version": "1.12"
-    },
-    "type": {
-      "instanceSchema": "title: kfp.Model\ntype: object\nproperties:\n  framework:\n    type: string\n  framework_version:\n    type: string\n"
-    }
-  }
